requirements.txt
python-dotenv==1.0.1
Flask==3.0.3
flask-smorest==0.44.0
SQLAlchemy==2.0.32
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-JWT-Extended==4.6.0
cryptography==43.0.1
flask-redis==0.4.0
mypy==1.11.2
request-id-flask==0.0.4
IP2Location==8.10.4
